backend/pyproject.toml
[project]
name = "umat-staff-chatbot-backend"
version = "0.1.0"
description = "FastAPI backend for the UMaT staff chatbot and admin console"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.16.5",
  "datasets>=3.1.0",
  "email-validator>=2.3.0",
  "evaluate>=0.4.3",
  "fastapi>=0.116.1",
  "pandas>=2.2.3",
  "passlib>=1.7.4",
  "pydantic-settings>=2.10.1",
  "python-jose[cryptography]>=3.5.0",
  "rapidfuzz>=3.14.1",
  "scikit-learn>=1.6.1",
  "spacy>=3.8.7",
  "sqlalchemy>=2.0.43",
  "sentence-transformers>=3.2.1",
  "torch>=2.6.0",
  "transformers>=4.48.0",
  "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1",
  "pytest>=8.4.2",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
